fix update price time

+++ coreV2/appdb/appdb.go
@@ -441,7 +441,7 @@ func (appDB *AppDB) Emission() (emission *big.Int) {
 }
 
 type TimePrice struct {
-	T      int64
+	T      uint64
 	R0, R1 *big.Int
 	Off    bool
 	Last   *big.Int
@@ -493,7 +493,7 @@ func (appDB *AppDB) SetPrice(t time.Time, r0, r1 *big.Int, lastReward *big.Int, 
 	defer appDB.mu.Unlock()
 
 	appDB.price = &TimePrice{
-		T:    t.UnixNano(),
+		T:    uint64(t.UnixNano()),
 		R0:   big.NewInt(0).Set(r0), // BIP
 		R1:   big.NewInt(0).Set(r1), // USDTE
 		Off:  off,
@@ -542,5 +542,5 @@ func (appDB *AppDB) GetPrice() (t time.Time, r0, r1 *big.Int, lastReward *big.In
 			panic(err)
 		}
 	}
-	return time.Unix(0, appDB.price.T), appDB.price.R0, appDB.price.R1, new(big.Int).Set(appDB.price.Last), appDB.price.Off
+	return time.Unix(0, int64(appDB.price.T)), appDB.price.R0, appDB.price.R1, new(big.Int).Set(appDB.price.Last), appDB.price.Off
 }
